Spam Detector
This is a very basic and simple project I began with which is a spam detector using Naive Bayes classifier and Adaboost classifier. The data set used is pre-processed and the features have already been extracted I just loaded the data set and carried out these different classification and compared them to see which gives the highest accuracy.